requirements.txt
pytest~=7.2.0
PyYAML~=6.0
numpy~=1.24.1
opencv-python~=4.7.0.68
